Add files via upload
Spam Mail Detection Project
This project focuses on building and training a machine learning model to detect and classify spam emails. The model uses a dataset of labeled emails to distinguish between spam and legitimate (ham) messages.   this project aims to provide an efficient and accurate solution for spam detection.